MotionMark/unit-tests/test-regression.html

<script src="../extensions/math.js"></script>
<script src="../extensions/point.js"></script>
<script src="../extensions/regression.js"></script>
<script src="resources/js-test.js"></script>
<body>
    <script>
        let points = [
            new Point(35, 17),
            new Point(85, 16),
            new Point(138, 17),
            new Point(191, 17),
            new Point(241, 16),
            new Point(294, 17),
            new Point(347, 17),
            new Point(397, 16),
            new Point(451, 17),
            new Point(501, 16),
            new Point(554, 17),
            new Point(607, 17),
            new Point(660, 17),
            new Point(710, 16),
            new Point(763, 17),
            new Point(816, 17),
            new Point(866, 16),
            new Point(919, 17),
            new Point(973, 17),
            new Point(1023, 16),
            new Point(1076, 17),
            new Point(1129, 17),
            new Point(1179, 16),
            new Point(1232, 17),
            new Point(1285, 17),
            new Point(1335, 16),
            new Point(1388, 17),
            new Point(1441, 17),
            new Point(1491, 16),
            new Point(1545, 17),
            new Point(1598, 17),
            new Point(1648, 16),
            new Point(1701, 17),
            new Point(1754, 17),
            new Point(1804, 16),
            new Point(1857, 17),
            new Point(1910, 17),
            new Point(1960, 16),
            new Point(2013, 17),
            new Point(2066, 17),
            new Point(2117, 16),
            new Point(2170, 17),
            new Point(2223, 17),
            new Point(2273, 16),
            new Point(2326, 17),
            new Point(2379, 17),
            new Point(2429, 16),
            new Point(2482, 17),
            new Point(2535, 17),
            new Point(2585, 16),
            new Point(2638, 17),
            new Point(2692, 17),
            new Point(2742, 16),
            new Point(2795, 17),
            new Point(2848, 17),
            new Point(2895, 15),
            new Point(2945, 16),
            new Point(2995, 16),
            new Point(3045, 16),
            new Point(3098, 17),
            new Point(3151, 17),
            new Point(3204, 17),
            new Point(3257, 17),
            new Point(3310, 17),
            new Point(3367, 18),
            new Point(3423, 18),
            new Point(3479, 18),
            new Point(3539, 19),
            new Point(3598, 19),
            new Point(3657, 19),
            new Point(3717, 19),
            new Point(3779, 20),
            new Point(3842, 20),
            new Point(3904, 20),
            new Point(3967, 20),
            new Point(4032, 21),
            new Point(4101, 22),
            new Point(4167, 21),
            new Point(4236, 22),
            new Point(4304, 22),
            new Point(4376, 23),
            new Point(4448, 23),
            new Point(4520, 23),
            new Point(4592, 23),
            new Point(4667, 24),
            new Point(4742, 24),
            new Point(4817, 24),
            new Point(4895, 25),
            new Point(4976, 26),
            new Point(5058, 26),
            new Point(5139, 26),
            new Point(5223, 27),
            new Point(5308, 27),
            new Point(5392, 27),
            new Point(5480, 28),
            new Point(5567, 28),
            new Point(5658, 29),
            new Point(5748, 29),
            new Point(5842, 30),
            new Point(5939, 31),
            new Point(6033, 30),
            new Point(6130, 31),
            new Point(6230, 32),
            new Point(6330, 32),
            new Point(6433, 33),
            new Point(6539, 34),
            new Point(6645, 34),
            new Point(6755, 35),
            new Point(6864, 35),
            new Point(6974, 35),
            new Point(7086, 36),
            new Point(7202, 37),
            new Point(7321, 38),
            new Point(7442, 39),
            new Point(7564, 39),
            new Point(7686, 39),
            new Point(7811, 40),
            new Point(7939, 41),
            new Point(8074, 43),
            new Point(8205, 42),
            new Point(8336, 42),
            new Point(8440, 44),
            new Point(8440, 33),
            new Point(8440, 34),
            new Point(8440, 33),
            new Point(8440, 34),
            new Point(8440, 33),
            new Point(8440, 33),
        ]
        let regression = new Regression(points);
        debug("regression.score = " + regression.score);
        shouldBeTrue("regression.score >= 2800 && regression.score <= 3500");

        let points2 = [
            new Point(52, 17),
            new Point(104, 17),
            new Point(155, 17),
            new Point(204, 16),
            new Point(258, 18),
            new Point(304, 15),
            new Point(355, 17),
            new Point(407, 17),
            new Point(455, 16),
            new Point(507, 17),
            new Point(558, 17),
            new Point(607, 16),
            new Point(658, 17),
            new Point(710, 17),
            new Point(758, 16),
            new Point(810, 17),
            new Point(862, 17),
            new Point(910, 16),
            new Point(962, 17),
            new Point(1013, 17),
            new Point(1062, 16),
            new Point(1113, 17),
            new Point(1165, 17),
            new Point(1213, 16),
            new Point(1265, 17),
            new Point(1316, 17),
            new Point(1365, 16),
            new Point(1416, 17),
            new Point(1468, 17),
            new Point(1516, 16),
            new Point(1568, 17),
            new Point(1620, 17),
            new Point(1668, 16),
            new Point(1720, 17),
            new Point(1771, 17),
            new Point(1820, 16),
            new Point(1871, 17),
            new Point(1923, 17),
            new Point(1971, 16),
            new Point(2023, 17),
            new Point(2074, 17),
            new Point(2123, 16),
            new Point(2174, 17),
            new Point(2226, 17),
            new Point(2274, 16),
            new Point(2326, 17),
            new Point(2377, 17),
            new Point(2429, 17),
            new Point(2478, 16),
            new Point(2526, 16),
            new Point(2572, 15),
            new Point(2617, 15),
            new Point(2662, 15),
            new Point(2711, 16),
            new Point(2759, 16),
            new Point(2805, 15),
            new Point(2853, 16),
            new Point(2902, 16),
            new Point(2954, 17),
            new Point(3002, 16),
            new Point(3054, 17),
            new Point(3105, 17),
            new Point(3160, 18),
            new Point(3214, 18),
            new Point(3269, 18),
            new Point(3323, 18),
            new Point(3378, 18),
            new Point(3436, 19),
            new Point(3493, 19),
            new Point(3551, 19),
            new Point(3611, 20),
            new Point(3672, 20),
            new Point(3733, 20),
            new Point(3796, 21),
            new Point(3857, 20),
            new Point(3921, 21),
            new Point(3984, 21),
            new Point(4054, 23),
            new Point(4121, 22),
            new Point(4187, 22),
            new Point(4257, 23),
            new Point(4327, 23),
            new Point(4397, 23),
            new Point(4466, 23),
            new Point(4539, 24),
            new Point(4612, 24),
            new Point(4688, 25),
            new Point(4764, 25),
            new Point(4842, 26),
            new Point(4924, 27),
            new Point(5003, 26),
            new Point(5082, 26),
            new Point(5164, 27),
            new Point(5246, 27),
            new Point(5330, 28),
            new Point(5415, 28),
            new Point(5503, 29),
            new Point(5594, 30),
            new Point(5685, 30),
            new Point(5776, 30),
            new Point(5867, 30),
            new Point(5961, 31),
            new Point(6058, 32),
            new Point(6158, 33),
            new Point(6258, 33),
            new Point(6361, 34),
            new Point(6464, 34),
            new Point(6567, 34),
            new Point(6671, 34),
            new Point(6777, 35),
            new Point(6889, 37),
            new Point(7001, 37),
            new Point(7116, 38),
            new Point(7231, 38),
            new Point(7347, 38),
            new Point(7465, 39),
            new Point(7589, 41),
            new Point(7714, 41),
            new Point(7841, 42),
            new Point(7965, 41),
            new Point(8089, 41),
            new Point(8187, 167),
            new Point(8187, 33),
            new Point(8187, 32)
        ];

        let regression2 = new Regression(points2);
        debug("regression2.score = " + regression2.score);
        shouldBeTrue("regression2.score >= 2800 && regression2.score <= 3500");
    </script>
</body>
